refactor: extract GeoLensSurfOps mixin from optim.py and geolens.py
Move aspheric surface management (add_aspheric, increase_aspheric_order
and helpers) from GeoLensOptim and surface pruning/shape correction
(prune_surf, correct_shape) from GeoLens into a dedicated
GeoLensSurfOps mixin, keeping GeoLensOptim focused on loss functions
and gradient-based optimisation.
